feat(structure): Support src/ directory distinct from app/
app/ should contain a demo app that shows off the library
features. If the package is meant to be a website, then app/
should just contain the source of the website itself. The
contents of app are assumed *not* to be reusable.

src/ should contain any reusable modules. For purposes
of building the local app, these modules are assumed
to be located in the namespace specified by
volo.namespace in package.json.

Closes #3